+++ src/main/java/com/sparcs/teamf/api/auth/config/SecurityConfig.java
@@ -43,7 +43,7 @@ public class SecurityConfig {
                 .cors().configurationSource(this::corsConfiguration).and()
                 .authorizeRequests()
                 .antMatchers("/v1/signup", "/v1/auth/login", "/v1/auth/email/**", "/v1/auth/refresh",
-                        "/v1/member/password-reset/email")
+                        "/v1/auth/password-reset/**")
                 .permitAll()
                 .anyRequest().authenticated().and()
                 .addFilterBefore(new JwtFilter(tokenProvider), UsernamePasswordAuthenticationFilter.class)
